requirements.txt
attrs==21.2.0
Authlib==0.15.5
bleach==3.3.1
bleach-allowlist==1.0.3
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
cryptography==36.0.0
dnspython==2.1.0
email-validator==1.1.3
feedwerk==1.0.0
Flask==1.1.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.1.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Markdown==3.3.6
MarkupSafe==2.0.1
mdx-breakless-lists==1.0.1
-e git://github.com/uniphil/github-links.git@d6a74287d8bc3b5046cf96a9324a438337edcdae#egg=mdx_gh_links
newrelic==6.2.0.156
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pygit2==1.6.1
Pygments==2.8.1
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
rauth==0.7.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.28
toml==0.10.2
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
